# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RESTful API for booking vehicle service appointments at dealerships.",
        "title": "Keyloop Appointment Booking API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/api/v1/appointments": {
            "get": {
                "description": "Returns appointments filtered by customer_id or dealership_id (one is required)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "appointments"
                ],
                "summary": "List appointments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Customer UUID",
                        "name": "customer_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Dealership UUID",
                        "name": "dealership_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpHandler.listAppointmentsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpHandler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpHandler.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new service appointment at a dealership for the chosen time slot",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "appointments"
                ],
                "summary": "Book an appointment",
                "parameters": [
                    {
                        "description": "Appointment details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httpHandler.bookAppointmentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/httpHandler.bookAppointmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpHandler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpHandler.errorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/httpHandler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpHandler.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/appointments/available-slots": {
            "get": {
                "description": "Returns available 30-minute time slots within 7 days of desired_time where a bay and qualified technician are both free",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "appointments"
                ],
                "summary": "List available slots",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dealership UUID",
                        "name": "dealership_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Vehicle UUID",
                        "name": "vehicle_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Customer UUID",
                        "name": "customer_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Service type identifiers (repeatable)",
                        "name": "services",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Desired start datetime (RFC3339)",
                        "name": "desired_time",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpHandler.availableSlotsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpHandler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpHandler.errorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/httpHandler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpHandler.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "httpHandler.appointmentJSON": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-01-10T08:00:00Z"
                },
                "customer_id": {
                    "type": "string",
                    "example": "c1a2b3c4-d5e6-7890-abcd-ef1234567890"
                },
                "dealership_id": {
                    "type": "string",
                    "example": "d1e2f3a4-b5c6-7890-abcd-ef1234567890"
                },
                "end_time": {
                    "type": "string",
                    "example": "2025-01-15T10:15:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "f1a2b3c4-d5e6-7890-abcd-ef1234567890"
                },
                "notes": {
                    "type": "string",
                    "example": ""
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httpHandler.serviceSnapshotJSON"
                    }
                },
                "start_time": {
                    "type": "string",
                    "example": "2025-01-15T09:00:00Z"
                },
                "status": {
                    "type": "string",
                    "example": "pending"
                },
                "vehicle_id": {
                    "type": "string",
                    "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                }
            }
        },
        "httpHandler.availableSlotsResponse": {
            "type": "object",
            "properties": {
                "available_slots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httpHandler.slotJSON"
                    }
                }
            }
        },
        "httpHandler.bookAppointmentRequest": {
            "type": "object",
            "required": [
                "customer_id",
                "dealership_id",
                "desired_start_time",
                "services",
                "vehicle_id"
            ],
            "properties": {
                "customer_id": {
                    "type": "string",
                    "example": "c1a2b3c4-d5e6-7890-abcd-ef1234567890"
                },
                "dealership_id": {
                    "type": "string",
                    "example": "d1e2f3a4-b5c6-7890-abcd-ef1234567890"
                },
                "desired_start_time": {
                    "type": "string",
                    "example": "2025-01-15T09:00:00Z"
                },
                "services": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "oil_change",
                        "tire_rotation"
                    ]
                },
                "vehicle_id": {
                    "type": "string",
                    "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                }
            }
        },
        "httpHandler.bookAppointmentResponse": {
            "type": "object",
            "properties": {
                "appointment_id": {
                    "type": "string",
                    "example": "f1a2b3c4-d5e6-7890-abcd-ef1234567890"
                },
                "duration_minutes": {
                    "type": "integer",
                    "example": 75
                },
                "message": {
                    "type": "string",
                    "example": "Appointment booked successfully"
                }
            }
        },
        "httpHandler.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "dealership not found"
                }
            }
        },
        "httpHandler.listAppointmentsResponse": {
            "type": "object",
            "properties": {
                "appointments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httpHandler.appointmentJSON"
                    }
                }
            }
        },
        "httpHandler.serviceSnapshotJSON": {
            "type": "object",
            "properties": {
                "estimated_minutes": {
                    "type": "integer",
                    "example": 30
                },
                "name": {
                    "type": "string",
                    "example": "Oil Change"
                },
                "price": {
                    "type": "number",
                    "example": 49.99
                },
                "service_id": {
                    "type": "string",
                    "example": "oil_change"
                }
            }
        },
        "httpHandler.slotJSON": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "string",
                    "example": "2025-01-15T10:00:00Z"
                },
                "start": {
                    "type": "string",
                    "example": "2025-01-15T09:00:00Z"
                }
            }
        }
    }
}